+++ src/main/java/org/codegenerator/generator/methodsequencefinders/internal/ArrayMethodSequenceFinder.java
@@ -1,4 +1,4 @@
-package org.codegenerator.generator.methodsequencefinders;
+package org.codegenerator.generator.methodsequencefinders.internal;
 
 import org.codegenerator.generator.codegenerators.buildables.Buildable;
 import org.codegenerator.history.History;
@@ -24,16 +24,6 @@ public class ArrayMethodSequenceFinder implements MethodSequenceFinderInternal {
     }
 
     @Override
-    public History<Executable> findReflectionCalls(@NotNull Object object) {
-        return findCallsInternal(object);
-    }
-
-    @Override
-    public History<JcMethod> findJacoDBCalls(@NotNull Object object) {
-        return findCallsInternal(object);
-    }
-
-    @Override
     public List<Object> findReflectionCallsInternal(@NotNull Object object, @NotNull History<Executable> history) {
         return findCallsInternal(object, history);
     }
@@ -41,12 +31,6 @@ public class ArrayMethodSequenceFinder implements MethodSequenceFinderInternal {
     @Override
     public List<Object> findJacoDBCallsInternal(@NotNull Object object, @NotNull History<JcMethod> history) {
         return findCallsInternal(object, history);
-    }
-
-    private <T> @NotNull History<T> findCallsInternal(@NotNull Object object) {
-        History<T> history = new History<>();
-        history.put(object, new HistoryArray<>(object, Collections.emptyList()));
-        return history;
     }
 
     private <T> @NotNull List<Object> findCallsInternal(@NotNull Object object, @NotNull History<T> history) {
