+++ src/main/java/org/codegenerator/generator/methodsequencefinders/internal/NullMethodSequenceFinder.java
@@ -1,4 +1,4 @@
-package org.codegenerator.generator.methodsequencefinders;
+package org.codegenerator.generator.methodsequencefinders.internal;
 
 import org.codegenerator.generator.codegenerators.buildables.Buildable;
 import org.codegenerator.history.History;
@@ -24,16 +24,6 @@ public class NullMethodSequenceFinder implements MethodSequenceFinderInternal {
     }
 
     @Override
-    public History<Executable> findReflectionCalls(@Nullable Object object) {
-        return findCallsInternal(object);
-    }
-
-    @Override
-    public History<JcMethod> findJacoDBCalls(@Nullable Object object) {
-        return findCallsInternal(object);
-    }
-
-    @Override
     public List<Object> findReflectionCallsInternal(@Nullable Object object, @NotNull History<Executable> history) {
         return findCallsInternal(object, history);
     }
@@ -41,12 +31,6 @@ public class NullMethodSequenceFinder implements MethodSequenceFinderInternal {
     @Override
     public List<Object> findJacoDBCallsInternal(@Nullable Object object, @NotNull History<JcMethod> history) {
         return findCallsInternal(object, history);
-    }
-
-    private <T> @NotNull History<T> findCallsInternal(Object object) {
-        History<T> history = new History<>();
-        history.put(object, new HistoryObject<>(object, Collections.emptyList()));
-        return history;
     }
 
     private <T> @Unmodifiable @NotNull List<Object> findCallsInternal(Object object, @NotNull History<T> history) {
